feat: 디렉토리 이동 및 안쓰는 파일 제거
