--- src/main/java/ex/sample/global/jwt/JwtStatus.java
@@ -1,5 +1,0 @@
-package ex.sample.global.jwt;
-
-public enum JwtStatus {
-    VALID, EXPIRED, INVALID
-}
