+++ src/main/java/ex/sample/global/redis/RedisUtil.java
@@ -1,38 +1,61 @@
 package ex.sample.global.redis;
 
+import com.fasterxml.jackson.core.JsonProcessingException;
+import com.fasterxml.jackson.databind.ObjectMapper;
+import ex.sample.global.exception.GlobalException;
+import ex.sample.global.response.ResponseCode;
+import java.time.Duration;
+import java.util.Optional;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import org.springframework.data.redis.core.StringRedisTemplate;
 import org.springframework.stereotype.Component;
 
-@Slf4j(topic = "RedisUtil")
+@Slf4j
 @Component
 @RequiredArgsConstructor
 public class RedisUtil {
 
-    private final RefreshTokenRepository refreshTokenRepository;
+    private final ObjectMapper objectMapper;
+    private final StringRedisTemplate redisTemplate;
 
-    public boolean isUserLogout(String nickname) {
-        boolean isLogin = refreshTokenRepository.existsById(nickname);
-        return !isLogin;
+    public <T> Optional<T> get(String key, Class<T> clazz) {
+        return Optional.ofNullable(redisTemplate.opsForValue().get(key))
+            .map(json -> parseJson(json, clazz));
     }
 
-    public void setUserLogin(String nickname, String refreshToken) {
-        RefreshToken refreshTokenItem = RefreshToken.builder()
-            .nickname(nickname)
-            .refreshToken(refreshToken)
-            .build();
-
-        refreshTokenRepository.save(refreshTokenItem);
+    public <T> void set(String key, T value) {
+        set(key, value, null);
     }
 
-    public void setUserLogout(String nickname) {
-        refreshTokenRepository.findById(nickname)
-            .ifPresent(refreshTokenRepository::delete);
+    public <T> void set(String key, T value, Duration ttl) {
+        String json = toJson(value);
+        if (ttl != null) {
+            redisTemplate.opsForValue().set(key, json, ttl);
+        } else {
+            redisTemplate.opsForValue().set(key, json);
+        }
     }
 
-    public String getUserRefreshToken(String nickname) {
-        return refreshTokenRepository.findById(nickname)
-            .map(RefreshToken::getRefreshToken)
-            .orElse("");
+    public void delete(String key) {
+        redisTemplate.delete(key);
+    }
+
+    private <T> String toJson(T value) {
+        try {
+            return objectMapper.writeValueAsString(value);
+        } catch (JsonProcessingException e) {
+            log.error("Redis serialization error: {}", value, e);
+            throw new GlobalException(ResponseCode.SYSTEM_ERROR);
+        }
+    }
+
+    private <T> T parseJson(String json, Class<T> clazz) {
+        try {
+            return objectMapper.readValue(json, clazz);
+        } catch (JsonProcessingException e) {
+            log.error("Redis deserialization error: {}", json, e);
+            throw new GlobalException(ResponseCode.SYSTEM_ERROR);
+        }
     }
 }
