--- src/main/java/ex/sample/global/redis/RefreshTokenRepository.java
@@ -1,9 +1,0 @@
-package ex.sample.global.redis;
-
-import org.springframework.data.repository.CrudRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface RefreshTokenRepository extends CrudRepository<RefreshToken, String> {
-
-}
