+++ src/main/java/ex/sample/global/swagger/SwaggerConfig.java
@@ -1,6 +1,7 @@
 package ex.sample.global.swagger;
 
-import ex.sample.global.jwt.JwtUtil;
+import ex.sample.global.security.jwt.JwtBearerUtils;
+import ex.sample.global.security.jwt.JwtConfig;
 import io.swagger.v3.oas.models.Components;
 import io.swagger.v3.oas.models.OpenAPI;
 import io.swagger.v3.oas.models.info.Info;
@@ -26,7 +27,7 @@ public class SwaggerConfig {
         addPageableSchemas(components); // Pageable 스키마 추가
         addSecuritySchemes(components); // 인증 스키마 추가
 
-        SecurityRequirement accessTokenRequirement = new SecurityRequirement().addList(JwtUtil.ACCESS_TOKEN_HEADER);
+        SecurityRequirement accessTokenRequirement = new SecurityRequirement().addList(JwtConfig.ACCESS_TOKEN_HEADER);
 
         return new OpenAPI()
             .info(info())
@@ -47,15 +48,15 @@ public class SwaggerConfig {
     }
 
     private void addSecuritySchemes(Components components) {
-        components.addSecuritySchemes(JwtUtil.ACCESS_TOKEN_HEADER, accessTokenSecurityScheme());
+        components.addSecuritySchemes(JwtConfig.ACCESS_TOKEN_HEADER, accessTokenSecurityScheme());
     }
 
     private SecurityScheme accessTokenSecurityScheme() {
         return new SecurityScheme()
-            .name(JwtUtil.ACCESS_TOKEN_HEADER) // Authorization 을 키로 함
+            .name(JwtConfig.ACCESS_TOKEN_HEADER) // Authorization 을 키로 함
             .type(SecurityScheme.Type.APIKEY) // name 값을 키로 함. (HTTP 설정 시 Authorization 고정이지만 확장 차)
             .in(In.HEADER) // 액세스 토큰은 헤더에 속함
-            .scheme(JwtUtil.BEARER_PREFIX) // Bearer 접두사 붙음
+            .scheme(JwtBearerUtils.BEARER_PREFIX) // Bearer 접두사 붙음
             .bearerFormat("JWT"); // 유형은 JWT
     }
